Improved instructions for ToIdentifierType flow
